otel: add support for HTTP server metrics
- add Response.Size
- log: Context.Length is int64 and use Response.Size instead of counting itself
- add Request.BodySize
- refactor internal/otel/semconv for shared attrs between span and metric

+++ log/common.go
@@ -67,7 +67,7 @@ func CommonLogFormatter(ctx *Context) (string, []slog.Attr) {
 		slog.String("uri", uri),
 		slog.String("proto", req.Proto),
 		slog.Int("status", ctx.Status),
-		slog.Int("length", ctx.Length),
+		slog.Int64("length", ctx.Length),
 	)
 
 	return message, []slog.Attr{details}
